third_party/WebKit/LayoutTests/http/tests/htmlimports/redirect.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="help" "https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#imported-document" />

<link id="same" rel="import"      href="http://127.0.0.1:8000/htmlimports/resources/hello.html">
<link id="cors" rel="import"      href="http://localhost:8000/htmlimports/resources/cors-basic.cgi">
<link id="xs"   rel="import"      href="http://localhost:8000/htmlimports/resources/hello.html">

<link id="same-same" rel="import" href="/security/resources/redir.php?url=http://127.0.0.1:8000/htmlimports/resources/hello.html">
<link id="same-cors" rel="import" href="/security/resources/redir.php?url=http://localhost:8000/htmlimports/resources/cors-basic.cgi">
<link id="same-xs"   rel="import" href="/security/resources/redir.php?url=http://localhost:8000/htmlimports/resources/hello.html">

</head>
<body>
<script>
function assert_loaded(id)
{
    var target = document.getElementById(id);
    assert_true(target.import instanceof Document)
    assert_true(0 <= target.import.querySelector("h1").innerHTML.indexOf("Hello"));
}

function assert_not_loaded(id)
{
    var target = document.getElementById(id);
    assert_equals(target.import, null);
}

test(function() { assert_loaded("same"); }, "Accessing to a same origin import");
test(function() { assert_loaded("cors"); }, "Accessing to CORS-enabled origin import");
test(function() { assert_not_loaded("xs"); }, "Blocking access to a cross origin import");

test(function() { assert_loaded("same-same"); }, "Accessing to a same origin import including same domain redirect");
test(function() { assert_loaded("same-cors"); }, "Accessing to a CORS origin import including same domain redirect");
test(function() { assert_not_loaded("same-xs"); }, "Blocking access to a cross origin iamport including same domain redirect");

</script>
</body>
</html>
